requirements.txt
# Generated by analyzing Python source files
#
# List of third-party libraries required for the project.
# To install them, run: pip install -r requirements.txt

# Used for loading environment variables from .env files
# from download_cash_flow.py and download_history_flow.py
python-dotenv

# Futu Open API Python SDK
# from download_cash_flow.py and download_history_flow.py
futu-api

# Data analysis and manipulation tool
# Used in all Python files
pandas

# Used for parsing YAML configuration files
# from futu_cli/common/config.py
PyYAML

